rename library to match class
