+++ lib/src/html_input_stream.dart
