+++ lib/src/tokenizer.dart
@@ -3,7 +3,7 @@ library tokenizer;
 import 'dart:collection';
 import 'package:html/parser.dart' show HtmlParser;
 import 'constants.dart';
-import 'inputstream.dart';
+import 'html_input_stream.dart';
 import 'token.dart';
 import 'utils.dart';
